Lift cohorts and cohort state up to Dashboard
Refactor Dashboard, Overview, and Cohorts so that the list of cohorts
and currently selected cohort are managed in the Dashboard component
and passed as props into Cohorts and Overview.

This change will simplify implementing the sidebar which will need
access to share the cohort state with other descendants of Dashboard.